requirements.txt
Django>=1.8,<2.0
psycopg2
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
